refactor(pages): split dashboard modules
