+++ pages/dashboard/store/state.js
@@ -1,0 +1,94 @@
+import { createAutocompleteState } from '../js/autocomplete.js';
+import {
+  createDefaultDataOverview,
+  createEmptyEditForm,
+  createEmptyFeedEditForm,
+  createEmptyFeedFilters,
+  createEmptyPushHistoryFilters,
+  createEmptySubscriptionFilters,
+  createEmptyUserEditForm,
+  createEmptyUserFilters,
+} from './helpers.js';
+
+export function createDashboardInitialState() {
+  return {
+    subs: [],
+    filteredSubs: [],
+    stats: {
+      total_subscriptions: 0,
+      active_subscriptions: 0,
+      total_feeds: 0,
+      unique_users: 0,
+    },
+    loading: true,
+    subFilters: createEmptySubscriptionFilters(),
+    editMode: false,
+    selectedIds: [],
+    subPagination: { page: 1, pageSize: 20 },
+    autocomplete: createAutocompleteState(),
+    panelVisible: false,
+    panelExpanded: false,
+    panelMode: 'detail',
+    editForm: createEmptyEditForm(),
+    detailSub: {},
+    detailItems: [],
+    historyDetail: null,
+    exportPreview: null,
+    activeTab: 'subs',
+    users: [],
+    userFilters: createEmptyUserFilters(),
+    usersLoading: false,
+    userEditMode: false,
+    selectedUserIds: [],
+    feeds: [],
+    feedFilters: createEmptyFeedFilters(),
+    feedsLoading: false,
+    feedEditMode: false,
+    selectedFeedIds: [],
+    feedEditPanelVisible: false,
+    feedEditForm: createEmptyFeedEditForm(),
+    pluginSettingsLoading: false,
+    dataManagementLoading: false,
+    dataManagementOverview: createDefaultDataOverview(),
+    exportFiles: [],
+    subscriptionDefaults: {
+      interval: 10,
+      notify: true,
+      send_mode: '自动',
+      length_limit: 0,
+      display_author: '自动',
+      display_via: '自动',
+      display_title: '自动',
+      display_entry_tags: false,
+      style: 'auto',
+      display_media: true,
+    },
+    historyRetentionDays: 30,
+    pushHistory: [],
+    pushHistoryLoading: false,
+    pushHistoryFilter: createEmptyPushHistoryFilters(),
+    pushHistoryTotal: 0,
+    pushHistoryEditMode: false,
+    selectedPushHistoryIds: [],
+    routeKbLoading: false,
+    routeKbTask: null,
+    routeKbStatus: null,
+    handlersLoading: false,
+    handlersRegistry: [],
+    expandedHandlerName: '',
+    userEditPanelVisible: false,
+    userEditForm: createEmptyUserEditForm(),
+    pendingActions: {},
+    toast: { show: false, message: '', type: 'success' },
+    dialog: {
+      show: false,
+      title: '',
+      message: '',
+      optionLabel: '',
+      optionValue: false,
+      okText: '确定',
+      okClass: 'btn-danger',
+      resolve: null,
+    },
+  };
+}
